Enable Servo to function as an event processor and implement API connectivity checks
